move gillick competence features to verbal consent page
